+++ src/Application.php
@@ -8,14 +8,9 @@
 namespace Cspray\Labrador;
 
 use Amp\Promise;
-use Cspray\Labrador\Plugin\BootablePlugin;
-use Cspray\Labrador\Plugin\EventAwarePlugin;
-use Cspray\Labrador\Plugin\Pluggable;
-use Cspray\Labrador\Plugin\PluginDependentPlugin;
-use Cspray\Labrador\Plugin\InjectorAwarePlugin;
 use Throwable;
 
-interface Application extends BootablePlugin, EventAwarePlugin, PluginDependentPlugin, InjectorAwarePlugin {
+interface Application {
 
     /**
      * Perform whatever logic or operations your application requires; return a Promise that resolves when you app is